requirements.txt
flask==2.3.3
flask-socketio==5.3.4
flask-cors==4.0.0
pillow==10.1.0
gunicorn==21.2.0
eventlet==0.33.3